Class8&9&10 Context Management/pyproject.toml
[project]
name = "class8-context-management"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.6",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
]
